♻️ rename type namespace to types
